+++ lib/csjs.js
@@ -1,6 +1,5 @@
 'use strict';
 
-var extractExtends = require('./css-extract-extends');
 var isComposition = require('./composition').isComposition;
 var buildExports = require('./build-exports');
 var scopify = require('./scopify');
@@ -29,9 +28,9 @@ module.exports = function csjsHandler(strings) {
 
   var localClasses = without(scoped.classes, ignores);
   var localKeyframes = without(scoped.keyframes, ignores);
-  // var compositions = extracted.compositions;
+  var compositions = scoped.compositions;
 
-  var exports = buildExports(localClasses, localKeyframes, {});
+  var exports = buildExports(localClasses, localKeyframes, compositions);
 
   return Object.defineProperty(exports, cssKey, {
     enumerable: false,
