--- lib/css-extract-extends.js
@@ -1,51 +1,0 @@
-'use strict';
-
-var makeComposition = require('./composition').makeComposition;
-
-var regex = /\.([^\s]+)(\s+)(extends\s+)(\.[^{]+)/g;
-
-module.exports = function extractExtends(css, hashed) {
-  var found, matches = [];
-  while (found = regex.exec(css)) {
-    matches.unshift(found);
-  }
-
-  function extractCompositions(acc, match) {
-    var extendee = getClassName(match[1]);
-    var keyword = match[3];
-    var extended = match[4];
-
-    // remove from output css
-    var index = match.index + match[1].length + match[2].length;
-    var len = keyword.length + extended.length;
-    acc.css = acc.css.slice(0, index) + " " + acc.css.slice(index + len + 1);
-
-    var extendedClasses = splitter(extended);
-
-    extendedClasses.forEach(function(className) {
-      if (!acc.compositions[extendee]) {
-        acc.compositions[extendee] = {};
-      }
-      if (!acc.compositions[className]) {
-        acc.compositions[className] = {};
-      }
-      acc.compositions[extendee][className] = acc.compositions[className];
-    });
-    return acc;
-  }
-
-  return matches.reduce(extractCompositions, {
-    css: css,
-    compositions: {}
-  });
-
-};
-
-function splitter(match) {
-  return match.split(',').map(getClassName);
-}
-
-function getClassName(str) {
-  var trimmed = str.trim();
-  return trimmed[0] === '.' ? trimmed.substr(1) : trimmed;
-}
